Updated HexCore and unit tests
